Ramilll-moex_scanner_telegram_bot-master/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
yfinance==0.2.33
